10774: renaming of applicable reducer files
