+++ src/applications/edu-benefits/10203/edu-benefits-entry.jsx
@@ -4,7 +4,7 @@ import '../sass/edu-benefits.scss';
 import startApp from 'platform/startup';
 
 import routes from './routes';
-import reducer from './reducer';
+import reducer from './reducers';
 import manifest from './manifest.json';
 
 startApp({
