+++ src/applications/edu-benefits/10203/reducers/index.js
@@ -1,20 +1,8 @@
-const GET_REMAINING_ENTITLEMENT_SUCCESS = 'GET_REMAINING_ENTITLEMENT_SUCCESS';
-
-const initialState = {
-  remainingEntitlement: null,
-};
-
-function post911GIBStatus(state = initialState, action) {
-  if (action.type === GET_REMAINING_ENTITLEMENT_SUCCESS) {
-    return {
-      ...state,
-      remainingEntitlement: action.data.remainingEntitlement,
-    };
-  }
-
-  return state;
-}
+import formConfig from '../config/form';
+import { createSaveInProgressFormReducer } from 'platform/forms/save-in-progress/reducers';
+import { post911GIBStatus } from './post911GIBStatus';
 
 export default {
+  form: createSaveInProgressFormReducer(formConfig),
   post911GIBStatus,
 };
