+++ src/applications/edu-benefits/10203/reducers/post911GIBStatus.js
@@ -1,0 +1,15 @@
+const GET_REMAINING_ENTITLEMENT_SUCCESS = 'GET_REMAINING_ENTITLEMENT_SUCCESS';
+
+const initialState = {
+  remainingEntitlement: null,
+};
+
+export function post911GIBStatus(state = initialState, action) {
+  if (action.type === GET_REMAINING_ENTITLEMENT_SUCCESS) {
+    return {
+      ...state,
+      remainingEntitlement: action.data.remainingEntitlement,
+    };
+  }
+  return state;
+}
